・再度CRTを利用するように修正
