+++ ChCppLibrary/Direct3D9/ChMesh/ChSkinMesh9.h
@@ -24,9 +24,7 @@ namespace ChMesh
 				};
 
 				ChVec3_9 pos;
-#ifdef CRT
 				std::vector<ChPtr::Shared<BonePow>>updateMat;
-#endif
 			};
 
 			struct Bone
@@ -34,10 +32,8 @@ namespace ChMesh
 				ChMat_9 baseMat;
 				ChMat_9 updateMat;
 				ChMat_9 offMat;
-#ifdef CRT
 				std::string myName;
 				ChPtr::Shared<Bone> offsetBone = nullptr;
-#endif
 			};
 
 			bool startPlayAniCheck = false;
@@ -53,8 +49,6 @@ namespace ChMesh
 			void Release();
 
 		public://Set Functions//
-
-#ifdef CRT
 
 			//XFileよりアニメーションを取得//
 			void SetAnimation(
@@ -106,7 +100,6 @@ namespace ChMesh
 					anis.second->SetOneFrameTime(_playMaxTime);
 				}
 			}
-#endif
 
 			void SetSkin()override;
 
@@ -115,8 +108,6 @@ namespace ChMesh
 			void SetOffsetVertex()override;
 
 		public://Get Functions//
-
-#ifdef CRT
 
 			inline std::vector<ChPtr::Shared<std::string>> GetAniNameList()
 			{
@@ -138,14 +129,10 @@ namespace ChMesh
 				return boneList[_str]->baseMat;
 			}
 
-#endif
-
 			friend MeshManager9;
 			friend MeshList9;
 
 		protected://Member Value//
-
-#ifdef CRT
 
 			std::string testName = "";
 
@@ -154,175 +141,16 @@ namespace ChMesh
 			//第一にアニメーション名//
 			//第二にBone名を入れる。//
 			using BoneAnimation = std::map<std::string, ChPtr::Shared<ChAnimationObject9>>;
-			
+
 			std::map<std::string, BoneAnimation> animations;
 
 			std::string nowPlayAniName = "";
 
 			std::map<std::string, ChPtr::Shared<Bone>> boneList;
-			
+
 			std::vector<std::string> boneNameList;
-#endif
 		};
 	
 }
-
-#ifdef CRT
-
-
-void ChMesh::SkinMesh9::Release()
-{
-	boneList.clear();
-	boneNameList.clear();
-	tAni.Release();
-}
-
-void ChMesh::SkinMesh9::SetAnimation(
-	const std::string& _animationName,
-	const std::string& _XFileName)
-{
-	BoneAnimation tmpAni;
-
-	tmpAni = ChANiSupport().CreateKeyFrame(_XFileName);
-
-	size_t aniNum = 0;
-	for (auto&& bones : boneList)
-	{
-		if (tmpAni.find(bones.first) == tmpAni.end())continue;
-		aniNum = tmpAni[bones.first]->GetAniCnt();
-		break;
-	}
-
-	for (auto&& bones : boneList)
-	{
-		if (tmpAni.find(bones.first) != tmpAni.end())continue;
-		auto ani = ChPtr::Make_S<ChAnimationObject9>();
-
-		for (size_t i = 0; i < aniNum; i++)
-		{
-			ani->SetAniObject(ChMat_9());
-		}
-		tmpAni[bones.first] = ani;
-
-	}
-
-	if (tmpAni.size() < boneList.size())return;
-
-	animations[_animationName] = tmpAni;
-
-	if (startPlayAniCheck)return;
-	startPlayAniCheck = true;
-	nowPlayAniName = _animationName;
-
-	for (auto&& ani : animations[nowPlayAniName])
-	{
-		ani.second->Play();
-	}
-}
-
-void ChMesh::SkinMesh9::SetSkin()
-{
-	if (ChPtr::NullCheck(mesh))return;
-	if (animations.size() <= 0)return;
-	if (boneList.size() <= 0)return;
-
-	MeshVertex9* tmpVer = nullptr;
-	mesh->LockVertexBuffer(NULL, (LPVOID*)&tmpVer);
-
-	if (ChPtr::NullCheck(tmpVer))return;
-
-	//BoneUpdate//
-	for (auto&& boneName : boneNameList)
-	{
-		ChMat_9 tmpMat = boneList[boneName]->offMat;
-
-		boneList[boneName]->updateMat = animations[nowPlayAniName][boneName]->Update();
-	}
-
-	//LastUpdateBone//
-
-	for (unsigned long i = boneNameList.size() - 1; i + 1 > 0; i--)
-	{
-
-		ChMat_9 tmpMat = boneList[boneNameList[i]]->offMat;
-
-		if (boneList[boneNameList[i]]->offsetBone == nullptr)
-		{
-
-			boneList[boneNameList[i]]->updateMat
-				= tmpMat * boneList[boneNameList[i]]->updateMat;
-
-			continue;
-		}
-
-		boneList[boneNameList[i]]->updateMat
-			= tmpMat
-			* boneList[boneNameList[i]]->updateMat
-			* boneList[boneNameList[i]]->offsetBone->updateMat;
-	}
-
-	//UpdateVertex//
-	for (unsigned long ver = 0; ver < mesh->GetNumVertices(); ver++)
-	{
-		ChMat_9 tmpMat;
-		tmpMat.Clear0();
-		ChVec3_9 tmpVec;
-
-		tmpVec = GetOffsetVertex(ver);
-
-		for (auto&& bones : boneVertexList[ver]->updateMat)
-		{
-
-			if (bones->waitPow <= 0.0f)continue;
-
-			ChMat_9 tmp;
-			tmp = (*bones->updateMat
-				* bones->waitPow);
-
-			tmpMat += tmp;
-
-		}
-
-		tmpVec.MatPos(tmpMat, tmpVec);
-
-		(tmpVer + ver)->pos = tmpVec;
-	}
-
-	mesh->UnlockVertexBuffer();
-
-}
-
-
-void ChMesh::SkinMesh9::SetOffsetVertex()
-{
-	if (ChPtr::NullCheck(mesh))return;
-
-	MeshVertex9* tmpVer = nullptr;
-	mesh->LockVertexBuffer(NULL, (LPVOID*)&tmpVer);
-
-	if (ChPtr::NullCheck(tmpVer))return;
-
-	for (unsigned long ver = 0; ver < mesh->GetNumVertices(); ver++)
-	{
-		auto tmpPos = ChPtr::Make_S<ChVec3_9>();
-		auto tmpVertex = ChPtr::Make_S<BoneVertex>();
-
-		*tmpPos = (tmpVer + ver)->pos;
-
-		tmpVertex->pos = (tmpVer + ver)->pos;
-
-		offsetVertexList.push_back(tmpPos);
-
-		boneVertexList.push_back(tmpVertex);
-	}
-
-
-	mesh->UnlockVertexBuffer();
-	return;
-}
-
-#endif
-
-#include"ChMeshShared9.h"
 
 #endif
